+++ src/main/java/team9499/commitbody/domain/exercise/service/ElasticExerciseService.java
@@ -1,8 +1,11 @@
 package team9499.commitbody.domain.exercise.service;
 
 import team9499.commitbody.domain.exercise.dto.CustomExerciseDto;
+import team9499.commitbody.domain.exercise.dto.SearchExerciseResponse;
 
 public interface ElasticExerciseService {
+
+    SearchExerciseResponse searchExercise(String name, String target, String equipment, Integer from, Integer size, Boolean like, String memberId, String exerciseType);
 
     void saveExercise(CustomExerciseDto customExerciseDto);
