+++ src/main/java/team9499/commitbody/domain/exercise/service/ExerciseService.java
@@ -9,8 +9,6 @@ import team9499.commitbody.domain.exercise.dto.response.ExerciseResponse;
 
 public interface ExerciseService {
 
-    SearchExerciseResponse searchExercise(String name, String target,String equipment, Integer from, Integer size, Boolean like, String memberId,String exerciseType);
-
     CustomExerciseDto saveCustomExercise(String exerciseName, ExerciseTarget exerciseTarget, ExerciseEquipment exerciseEquipment, Long memberId, MultipartFile file);
 
     CustomExerciseDto updateCustomExercise(String exerciseName, ExerciseTarget exerciseTarget, ExerciseEquipment exerciseEquipment, Long memberId, Long customExerciseId,MultipartFile file);
